chore: declare method names for search module
