--- hr-finder/src/HRmanager0228/cli/controller/SalaryController.java
@@ -1,4 +1,0 @@
-package HRmanager0228.cli.controller;
-
-public class SalaryController {
-}
