--- hr-finder/src/HRmanager0228/cli/io/EmployeeIO.java
@@ -1,4 +1,0 @@
-package HRmanager0228.cli.io;
-
-public class EmployeeIO {
-}
