--- hr-finder/src/HRmanager0228/cli/io/SalaryIO.java
@@ -1,4 +1,0 @@
-package HRmanager0228.cli.io;
-
-public class SalaryIO {
-}
