--- hr-finder/src/HRmanager0228/config/configdemo.java
@@ -1,4 +1,0 @@
-package HRmanager0228.config;
-// db나 파일 연결,로드 등 관련 환경 통합관리
-public class configdemo {
-}
