--- hr-finder/src/HRmanager0228/service/SalaryServiceImpl.java
@@ -1,4 +1,0 @@
-package HRmanager0228.service;
-
-public class SalaryServiceImpl {
-}
