--- hr-finder/src/HRmanager0228/util/validation/Validator.java
@@ -1,8 +1,0 @@
-package HRmanager0228.util.validation;
-
-import java.util.function.Function;
-import java.util.function.Predicate;
-
-public interface Validator {
-    <T> T readValidated(String prompt, Function<String, T> parser, Predicate<T> validator, String errorMessage);
-}
