+++ hr-finder/src/HRmanager0301/cli/controller/EmployeeController.java
@@ -1,0 +1,4 @@
+package HRmanager0301.cli.controller;
+
+public class EmployeeController {
+}
