+++ hr-finder/src/HRmanager0301/cli/io/SalaryIO.java
@@ -1,0 +1,4 @@
+package HRmanager0301.cli.io;
+
+public class SalaryIO {
+}
