+++ hr-finder/src/HRmanager0301/dao/EmployeeDaoImpl.java
@@ -1,0 +1,4 @@
+package HRmanager0301.dao;
+
+public class EmployeeDaoImpl {
+}
