+++ hr-finder/src/HRmanager0301/dao/SalaryDao.java
@@ -1,0 +1,4 @@
+package HRmanager0301.dao;
+
+public interface SalaryDao {
+}
