+++ hr-finder/src/HRmanager0301/dao/SalaryDaoImpl.java
@@ -1,0 +1,4 @@
+package HRmanager0301.dao;
+
+public class SalaryDaoImpl {
+}
