+++ hr-finder/src/HRmanager0301/service/EmployeeService.java
@@ -1,0 +1,4 @@
+package HRmanager0301.service;
+
+public interface EmployeeService {
+}
