+++ hr-finder/src/HRmanager0301/util/utildemo.java
@@ -1,0 +1,4 @@
+package HRmanager0301.util;
+// validator같은 공통된 기능을 제공하는.. 예외처리도 ?
+public class utildemo {
+}
